more module reworks
