+++ src/lib.rs
@@ -13,9 +13,9 @@ pub mod error;
 pub mod github;
 pub mod bot;
 pub mod core;
+pub mod git_ops;
 pub mod gitlab;
 pub mod merge_request;
-pub mod rebase;
 pub mod server;
 pub mod types;
 pub mod vanity_service;
